Rewrite movingRate so updates are instantly reflected in the calculated rate.
Otherwise either all retries would be permitted or forbidden for a second,
until the new data is merged into the history. This was bad because it
causes batching of retries, exactly what we're trying to avoid.

The new code also allows for arbitrary bucket lengths, i.e. granularity
other than 1 second.